+++ Sources/NSMessagePack/ConvenienceProperties.swift
@@ -49,9 +49,9 @@ extension MessagePackValue {
     @available(*, deprecated, message: "use int64Value: instead")
     public var integerValue: Int64? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return value
-        case .uint(let value) where value <= UInt64(Int64.max):
+        case .uint64(let value) where value <= UInt64(Int64.max):
             return Int64(value)
         default:
             return nil
@@ -62,9 +62,9 @@ extension MessagePackValue {
     /// appropriately valued `.uint`, `nil` otherwise.
     public var intValue: Int? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return Int(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return Int(exactly: value)
         default:
             return nil
@@ -75,9 +75,9 @@ extension MessagePackValue {
     /// `.uint`, `nil` otherwise.
     public var int8Value: Int8? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return Int8(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return Int8(exactly: value)
         default:
             return nil
@@ -88,9 +88,9 @@ extension MessagePackValue {
     /// `.uint`, `nil` otherwise.
     public var int16Value: Int16? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return Int16(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return Int16(exactly: value)
         default:
             return nil
@@ -101,9 +101,9 @@ extension MessagePackValue {
     /// `.uint`, `nil` otherwise.
     public var int32Value: Int32? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return Int32(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return Int32(exactly: value)
         default:
             return nil
@@ -114,9 +114,9 @@ extension MessagePackValue {
     /// `.uint`, `nil` otherwise.
     public var int64Value: Int64? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return value
-        case .uint(let value):
+        case .uint64(let value):
             return Int64(exactly: value)
         default:
             return nil
@@ -129,9 +129,9 @@ extension MessagePackValue {
     @available(*, deprecated, message: "use uint64Value: instead")
     public var unsignedIntegerValue: UInt64? {
         switch self {
-        case .int(let value) where value >= 0:
+        case .int64(let value) where value >= 0:
             return UInt64(value)
-        case .uint(let value):
+        case .uint64(let value):
             return value
         default:
             return nil
@@ -142,9 +142,9 @@ extension MessagePackValue {
     /// appropriately valued `.int`, `nil` otherwise.
     public var uintValue: UInt? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return UInt(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return UInt(exactly: value)
         default:
             return nil
@@ -155,9 +155,9 @@ extension MessagePackValue {
     /// `.int`, `nil` otherwise.
     public var uint8Value: UInt8? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return UInt8(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return UInt8(exactly: value)
         default:
             return nil
@@ -168,9 +168,9 @@ extension MessagePackValue {
     /// `.int`, `nil` otherwise.
     public var uint16Value: UInt16? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return UInt16(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return UInt16(exactly: value)
         default:
             return nil
@@ -181,9 +181,9 @@ extension MessagePackValue {
     /// `.int`, `nil` otherwise.
     public var uint32Value: UInt32? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return UInt32(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return UInt32(exactly: value)
         default:
             return nil
@@ -194,9 +194,9 @@ extension MessagePackValue {
     /// `.int`, `nil` otherwise.
     public var uint64Value: UInt64? {
         switch self {
-        case .int(let value):
+        case .int64(let value):
             return UInt64(exactly: value)
-        case .uint(let value):
+        case .uint64(let value):
             return value
         default:
             return nil
